+++ WebAPI/Controllers/JobsController.cs
@@ -12,21 +12,37 @@ namespace WebAPI.Controllers
     {
         MyContext context = new MyContext();
 
+        public bool CheckTok()
+        {
+            return context.Tokens.Select(token => token.text).Contains(this.Request.Headers.GetValues("tok").First());
+        }
+
+
         // GET: api/Jobs
         public IEnumerable<Jobs> Get()
         {
+            if (!CheckTok())
+                return null;
+
+
             return context.Jobs;
         }
 
         // GET: api/Jobs/5
         public Jobs Get(int id)
         {
+            if (!CheckTok())
+                return null;
+
             return context.Jobs.Find(id);
         }
 
         // POST: api/Jobs
         public void Post([FromBody]Jobs jobs)
         {
+            if (!CheckTok())
+                return;
+
 
             this.context.Jobs.Add(jobs);
             this.context.SaveChanges();
@@ -37,6 +53,10 @@ namespace WebAPI.Controllers
         // PUT: api/Jobs/5
         public void Put(int id, [FromBody]Jobs jobs)
         {
+            if (!CheckTok())
+                return;
+
+
             Jobs current = this.context.Jobs.Find(id);
 
             current.ID_Computer = jobs.ID_Computer;
@@ -51,6 +71,11 @@ namespace WebAPI.Controllers
         // DELETE: api/Jobs/5
         public void Delete(int id)
         {
+            if (!CheckTok())
+                return;
+                
+
+
             Jobs jobs = this.context.Jobs.Find(id);
 
             this.context.Jobs.Remove(jobs);
