+++ WebAPI/Controllers/TokensController.cs
@@ -5,7 +5,7 @@ using System.Net;
 using System.Net.Http;
 using System.Web.Http;
 using WebAPI.Models;
-
+using WebAPI.TOKENS;
 
 namespace WebAPI.Controllers
 {
@@ -13,21 +13,36 @@ namespace WebAPI.Controllers
     {
         MyContext context = new MyContext();
 
+
+        public bool CheckTok()
+        {
+            return context.Tokens.Select(token => token.text).Contains(this.Request.Headers.GetValues("tok").First());
+        }
+
         // GET: api/Tokens
         public IEnumerable<Tokens> Get()
         {
+            if (!CheckTok())
+                return null;
+
             return context.Tokens;
         }
 
         // GET: api/Tokens/5
         public Tokens Get(int id)
         {
+            if (!CheckTok())
+                return null;
+
             return context.Tokens.Find(id);
         }
 
         // POST: api/Tokens
         public void Post([FromBody]Tokens token)
         {
+            if (!CheckTok())
+                return;
+
             this.context.Tokens.Add(token);
             this.context.SaveChanges();
         }
@@ -35,6 +50,9 @@ namespace WebAPI.Controllers
         // PUT: api/Tokens/5
         public void Put(int id, [FromBody]Tokens token)
         {
+            if (!CheckTok())
+                return;
+
             Tokens current = this.context.Tokens.Find(id);
 
             current.life = token.life;
@@ -46,10 +64,55 @@ namespace WebAPI.Controllers
         // DELETE: api/Tokens/5
         public void Delete(int id)
         {
+            if (!CheckTok())
+                return;
+
+
             Tokens token = this.context.Tokens.Find(id);
 
             this.context.Tokens.Remove(token);
             this.context.SaveChanges();
+        }
+
+        [HttpPost]
+        public string GetToken([FromBody]UsernamePassword uspas)
+        {
+       
+
+            User  d =   context.Users.FirstOrDefault(item => item.Username == uspas.Username);
+
+            if( d != null && d.Password == uspas.Password)
+            {
+                Tokens token = TokensGenerateVer.TK();
+
+                this.context.Tokens.Add(token);
+                this.context.SaveChanges();
+
+                return token.text;
+
+            }
+
+            else
+            {
+                return null;
+
+            }
+
+
+        }
+
+
+        public void DeleteToken([FromBody]string token)
+        {
+
+
+            Tokens tok = context.Tokens.FirstOrDefault(item => item.text == token);
+
+            if(tok != null)
+            {
+                Delete(tok.ID);
+
+            }
         }
     }
 }
